requirements.txt
amqp==2.4.1
beautifulsoup4==4.7.1
billiard==3.5.0.5
celery==4.2.1
celery-redbeat==0.12.0
cffi==1.12.2
dj-database-url==0.5.0
Django==2.0
django-crispy-forms==1.7.2
django-tables2==2.0.5
gcloud==0.17.0
gevent==1.4.0
googleapis-common-protos==1.5.8
greenlet==0.4.15
httplib2==0.12.1
jws==0.1.3
kombu==4.3.0
oauth2client==3.0.0
protobuf==3.7.0rc3
psycopg2==2.7.7
psycopg2-binary==2.7.7
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycparser==2.19
pycryptodome==3.4.3
Pyrebase==3.0.27
python-dateutil==2.8.0
python-jwt==2.0.1
pytz==2018.9
redis==2.10.6
requests==2.11.1
requests-toolbelt==0.7.0
rsa==4.0
six==1.12.0
soupsieve==1.8
tenacity==5.0.3
vine==1.2.0
